requirements.txt
numpy
flask
pillow
gunicorn
flask_cors
scikit-learn==0.22.1
joblib